+++ project/app/Http/Livewire/Agenda.php
@@ -3,6 +3,7 @@
 namespace App\Http\Livewire;
 
 use App\Http\Forms\ResourceForm;
+use App\Resource;
 use App\Task;
 use Carbon\Carbon;
 use Illuminate\Database\Eloquent\Collection;
@@ -13,14 +14,17 @@ class Agenda extends Component
     public string $date;
     public array $days;
     public ?string $resource_id = null;
+    public ?string $searchResource = null;
+    public $results;
+    public bool $showResources = false;
     private Carbon $week;
 
     public function render()
     {
         $tasks = $this->index();
-        $selectResources = ResourceForm::select();
+        $this->results = $this->searchResource();
 
-        return view('livewire.agenda', compact('tasks', 'selectResources'));
+        return view('livewire.agenda', compact('tasks'));
     }
 
     public function mount()
@@ -55,6 +59,10 @@ class Agenda extends Component
     public function index() : Collection
     {
         if ( empty($this->resource_id) ) {
+            if ( ! empty (session()->get('resource')) ) {
+                $this->resource_id = session()->get('resource');
+            }
+
             return Task::where('day', '>=', $this->days['Lundi'])
                        ->where('day', '<=', $this->days['Vendredi'])
                        ->orderBy('day')
@@ -63,6 +71,8 @@ class Agenda extends Component
                        ->orderBy('project_id')
                        ->get();
         }
+
+        session()->put('resource', $this->resource_id);
 
         return Task::where('resource_id', $this->resource_id)
                    ->where('day', '>=', $this->days['Lundi'])
@@ -86,5 +96,36 @@ class Agenda extends Component
             'Jeudi'    => $current->addDay()->format('Y-m-d'),
             'Vendredi' => $current->addDay()->format('Y-m-d'),
         ];
+    }
+
+    public function searchResource()
+    {
+        return Resource::where('first_name', 'like', "%{$this->searchResource}%")->get();
+    }
+
+    public function setResource(string $id)
+    {
+        $resource = Resource::find($id);
+        session()->put('resource', $id);
+        $this->resource_id = $id;
+        $this->searchResource = $resource->first_name;
+    }
+
+    public function refresh()
+    {
+        session()->put('resource', '');
+        $this->resource_id = null;
+        $this->searchResource = null;
+    }
+
+    public function showResources ()
+    {
+        $this->searchResource = null;
+        $this->showResources = true;
+    }
+
+    public function hideResources ()
+    {
+        $this->showResources = false;
     }
 }